+++ Tests/TuistKitTests/ProjectEditor/ProjectEditorMapperTests.swift
@@ -21,11 +21,13 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         super.tearDown()
     }
 
-    func test_edit_when_there_are_helpers() throws {
+    func test_edit_when_there_are_helpers_and_setup_and_config() throws {
         // Given
         let sourceRootPath = try temporaryPath()
         let xcodeProjPath = sourceRootPath.appending(component: "Project.xcodeproj")
         let manifestPaths = [sourceRootPath].map { $0.appending(component: "Project.swift") }
+        let setupPath = sourceRootPath.appending(component: "Setup.swift")
+        let configPath = sourceRootPath.appending(components: Constants.tuistDirectoryName, "Config.swift")
         let helperPaths = [sourceRootPath].map { $0.appending(component: "Project+Template.swift") }
         let templates = [sourceRootPath].map { $0.appending(component: "template") }
         let projectDescriptionPath = sourceRootPath.appending(component: "ProjectDescription.framework")
@@ -35,6 +37,8 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         let (project, graph) = try subject.map(tuistPath: tuistPath,
                                                sourceRootPath: sourceRootPath,
                                                xcodeProjPath: xcodeProjPath,
+                                               setupPath: setupPath,
+                                               configPath: configPath,
                                                manifests: manifestPaths,
                                                helpers: helperPaths,
                                                templates: templates,
@@ -42,8 +46,9 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
 
         // Then
         let targetNodes = graph.targets.values.lazy.flatMap { targets in targets.compactMap { $0 } }.sorted(by: { $0.target.name < $1.target.name })
-        XCTAssertEqual(targetNodes.count, 3)
-        XCTAssertEqual(targetNodes.last?.dependencies, Array(targetNodes.dropLast()))
+
+        XCTAssertEqual(targetNodes.count, 5)
+        XCTAssertEqual(Set(targetNodes.last?.dependencies ?? []), Set(targetNodes.dropLast()))
 
         // Generated Manifests target
         let manifestsTarget = try XCTUnwrap(project.targets.first)
@@ -54,11 +59,11 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         XCTAssertEqual(manifestsTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
         XCTAssertEqual(manifestsTarget.sources.map { $0.path }, manifestPaths)
         XCTAssertEqual(manifestsTarget.filesGroup, .group(name: "Manifests"))
-        XCTAssertEqual(manifestsTarget.dependencies, [.target(name: "ProjectDescriptionHelpers"), .target(name: "Templates")])
+        XCTAssertEqual(manifestsTarget.dependencies, [.target(name: "ProjectDescriptionHelpers")])
 
         // Generated Helpers target
         let helpersTarget = try XCTUnwrap(project.targets.last(where: { $0.name == "ProjectDescriptionHelpers" }))
-        XCTAssertEqual(targetNodes.dropLast().first?.target, helpersTarget)
+        XCTAssertEqual(targetNodes.dropFirst().first?.target, helpersTarget)
 
         XCTAssertEqual(helpersTarget.name, "ProjectDescriptionHelpers")
         XCTAssertEqual(helpersTarget.platform, .macOS)
@@ -66,7 +71,7 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         XCTAssertEqual(helpersTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
         XCTAssertEqual(helpersTarget.sources.map { $0.path }, helperPaths)
         XCTAssertEqual(helpersTarget.filesGroup, .group(name: "Manifests"))
-        XCTAssertEqual(helpersTarget.dependencies, [])
+        XCTAssertEmpty(helpersTarget.dependencies)
 
         // Generated Templates target
         let templatesTarget = try XCTUnwrap(project.targets.last(where: { $0.name == "Templates" }))
@@ -78,7 +83,31 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         XCTAssertEqual(templatesTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
         XCTAssertEqual(templatesTarget.sources.map { $0.path }, templates)
         XCTAssertEqual(templatesTarget.filesGroup, .group(name: "Manifests"))
-        XCTAssertEqual(templatesTarget.dependencies, [])
+        XCTAssertEmpty(templatesTarget.dependencies)
+
+        // Generated Setup target
+        let setupTarget = try XCTUnwrap(project.targets.last(where: { $0.name == "Setup" }))
+        XCTAssertEqual(targetNodes.dropLast().dropLast().last?.target, setupTarget)
+
+        XCTAssertEqual(setupTarget.name, "Setup")
+        XCTAssertEqual(setupTarget.platform, .macOS)
+        XCTAssertEqual(setupTarget.product, .staticFramework)
+        XCTAssertEqual(setupTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
+        XCTAssertEqual(setupTarget.sources.map { $0.path }, [setupPath])
+        XCTAssertEqual(setupTarget.filesGroup, .group(name: "Manifests"))
+        XCTAssertEmpty(setupTarget.dependencies)
+
+        // Generated Config target
+        let configTarget = try XCTUnwrap(project.targets.last(where: { $0.name == "Config" }))
+        XCTAssertEqual(targetNodes.first?.target, configTarget)
+
+        XCTAssertEqual(configTarget.name, "Config")
+        XCTAssertEqual(configTarget.platform, .macOS)
+        XCTAssertEqual(configTarget.product, .staticFramework)
+        XCTAssertEqual(configTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
+        XCTAssertEqual(configTarget.sources.map { $0.path }, [configPath])
+        XCTAssertEqual(configTarget.filesGroup, .group(name: "Manifests"))
+        XCTAssertEmpty(configTarget.dependencies)
 
         // Generated Project
         XCTAssertEqual(project.path, sourceRootPath)
@@ -103,7 +132,7 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         XCTAssertEqual(runAction.arguments, Arguments(launchArguments: [generateArgument: true]))
     }
 
-    func test_edit_when_there_are_no_helpers() throws {
+    func test_edit_when_there_are_no_helpers_and_no_setup_and_no_config() throws {
         // Given
         let sourceRootPath = try temporaryPath()
         let xcodeProjPath = sourceRootPath.appending(component: "Project.xcodeproj")
@@ -117,6 +146,8 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         let (project, graph) = try subject.map(tuistPath: tuistPath,
                                                sourceRootPath: sourceRootPath,
                                                xcodeProjPath: xcodeProjPath,
+                                               setupPath: nil,
+                                               configPath: nil,
                                                manifests: manifestPaths,
                                                helpers: helperPaths,
                                                templates: templates,
@@ -124,8 +155,9 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
 
         // Then
         let targetNodes = graph.targets.values.flatMap { targets in targets.compactMap { $0 } }.sorted(by: { $0.target.name < $1.target.name })
+
         XCTAssertEqual(targetNodes.count, 1)
-        XCTAssertEqual(targetNodes.first?.dependencies, [])
+        XCTAssertEmpty(targetNodes.first?.dependencies ?? [])
 
         // Generated Manifests target
         let manifestsTarget = try XCTUnwrap(project.targets.first)
@@ -136,7 +168,7 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         XCTAssertEqual(manifestsTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
         XCTAssertEqual(manifestsTarget.sources.map { $0.path }, manifestPaths)
         XCTAssertEqual(manifestsTarget.filesGroup, .group(name: "Manifests"))
-        XCTAssertEqual(manifestsTarget.dependencies, [])
+        XCTAssertEmpty(manifestsTarget.dependencies)
 
         // Generated Project
         XCTAssertEqual(project.path, sourceRootPath)
@@ -165,6 +197,8 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         // Given
         let sourceRootPath = try temporaryPath()
         let xcodeProjPath = sourceRootPath.appending(component: "Project.xcodeproj")
+        let setupPath = sourceRootPath.appending(component: "Setup.swift")
+        let configPath = sourceRootPath.appending(components: Constants.tuistDirectoryName, "Config.swift")
         let otherProjectPath = "Module"
         let manifestPaths = [
             sourceRootPath.appending(component: "Project.swift"),
@@ -179,6 +213,8 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         let (project, graph) = try subject.map(tuistPath: tuistPath,
                                                sourceRootPath: sourceRootPath,
                                                xcodeProjPath: xcodeProjPath,
+                                               setupPath: setupPath,
+                                               configPath: configPath,
                                                manifests: manifestPaths,
                                                helpers: helperPaths,
                                                templates: templates,
@@ -186,13 +222,15 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
 
         // Then
         let targetNodes = graph.targets.values.flatMap { targets in targets.compactMap { $0 } }.sorted(by: { $0.target.name < $1.target.name })
-        XCTAssertEqual(targetNodes.count, 2)
-        XCTAssertEqual(targetNodes.first?.dependencies, [])
-        XCTAssertEqual(targetNodes.last?.dependencies, [])
+
+        // expecting `targetNodes == [Config, ModuleManifests, Setup, TemporaryDirectory.XXXMainifest]`
+        XCTAssertEqual(targetNodes.count, 4)
+        XCTAssertEmpty(targetNodes.first?.dependencies ?? [])
+        XCTAssertEqual(Set(targetNodes.last?.dependencies ?? []), Set([targetNodes[0], targetNodes[2]]))
 
         // Generated Manifests target
         let manifestOneTarget = try XCTUnwrap(project.targets.first(where: { $0.name == "ModuleManifests" }))
-        XCTAssertEqual(targetNodes.first?.target, manifestOneTarget)
+        XCTAssertEqual(targetNodes.dropFirst().first?.target, manifestOneTarget)
 
         XCTAssertEqual(manifestOneTarget.name, "ModuleManifests")
         XCTAssertEqual(manifestOneTarget.platform, .macOS)
@@ -200,7 +238,7 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         XCTAssertEqual(manifestOneTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
         XCTAssertEqual(manifestOneTarget.sources.map { $0.path }, [manifestPaths.last])
         XCTAssertEqual(manifestOneTarget.filesGroup, .group(name: "Manifests"))
-        XCTAssertEqual(manifestOneTarget.dependencies, [])
+        XCTAssertEmpty(manifestOneTarget.dependencies)
 
         // Generated Manifests target
         let manifestTwoTarget = try XCTUnwrap(project.targets.first(where: { $0.name != "ModuleManifests" }))
@@ -211,7 +249,31 @@ final class ProjectEditorMapperTests: TuistUnitTestCase {
         XCTAssertEqual(manifestTwoTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
         XCTAssertEqual(manifestTwoTarget.sources.map { $0.path }, [manifestPaths.first])
         XCTAssertEqual(manifestTwoTarget.filesGroup, .group(name: "Manifests"))
-        XCTAssertEqual(manifestTwoTarget.dependencies, [])
+        XCTAssertEmpty(manifestTwoTarget.dependencies)
+
+        // Generated Setup target
+        let setupTarget = try XCTUnwrap(project.targets.last(where: { $0.name == "Setup" }))
+        XCTAssertEqual(targetNodes.dropLast().last?.target, setupTarget)
+
+        XCTAssertEqual(setupTarget.name, "Setup")
+        XCTAssertEqual(setupTarget.platform, .macOS)
+        XCTAssertEqual(setupTarget.product, .staticFramework)
+        XCTAssertEqual(setupTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
+        XCTAssertEqual(setupTarget.sources.map { $0.path }, [setupPath])
+        XCTAssertEqual(setupTarget.filesGroup, .group(name: "Manifests"))
+        XCTAssertEmpty(setupTarget.dependencies)
+
+        // Generated Config target
+        let configTarget = try XCTUnwrap(project.targets.last(where: { $0.name == "Config" }))
+        XCTAssertEqual(targetNodes.first?.target, configTarget)
+
+        XCTAssertEqual(configTarget.name, "Config")
+        XCTAssertEqual(configTarget.platform, .macOS)
+        XCTAssertEqual(configTarget.product, .staticFramework)
+        XCTAssertEqual(configTarget.settings, expectedSettings(sourceRootPath: sourceRootPath))
+        XCTAssertEqual(configTarget.sources.map { $0.path }, [configPath])
+        XCTAssertEqual(configTarget.filesGroup, .group(name: "Manifests"))
+        XCTAssertEmpty(configTarget.dependencies)
 
         // Generated Project
         XCTAssertEqual(project.path, sourceRootPath)