+++ Tests/TuistLoaderTests/Loaders/SetupLoaderTests.swift
@@ -21,17 +21,20 @@ final class SetupLoaderTests: TuistUnitTestCase {
     var subject: SetupLoader!
     var upLinter: MockUpLinter!
     var manifestLoader: MockManifestLoader!
+    var manifestFilesLocator: MockManifestFilesLocator!
 
     override func setUp() {
         super.setUp()
         upLinter = MockUpLinter()
         manifestLoader = MockManifestLoader()
-        subject = SetupLoader(upLinter: upLinter, manifestLoader: manifestLoader)
+        manifestFilesLocator = MockManifestFilesLocator()
+        subject = SetupLoader(upLinter: upLinter, manifestLoader: manifestLoader, manifestFilesLocator: manifestFilesLocator)
     }
 
     override func tearDown() {
         upLinter = nil
         manifestLoader = nil
+        manifestFilesLocator = nil
         subject = nil
         super.tearDown()
     }
@@ -39,7 +42,9 @@ final class SetupLoaderTests: TuistUnitTestCase {
     func test_meet_when_no_actions() throws {
         // given
         let projectPath = try temporaryPath()
-        try FileHandler.shared.touch(projectPath.appending(component: Manifest.setup.fileName(projectPath)))
+        let setupPath = projectPath.appending(component: Manifest.setup.fileName(projectPath))
+        manifestFilesLocator.locateSetupStub = setupPath
+
         var receivedPaths = [String]()
         manifestLoader.loadSetupStub = { gotPath in
             receivedPaths.append(gotPath.pathString)
@@ -56,7 +61,7 @@ final class SetupLoaderTests: TuistUnitTestCase {
     func test_meet_when_actions_provided() throws {
         // given
         let projectPath = try temporaryPath()
-        try FileHandler.shared.touch(projectPath.appending(component: Manifest.setup.fileName(projectPath)))
+        manifestFilesLocator.locateSetupStub = projectPath.appending(component: Manifest.setup.fileName(projectPath))
 
         let mockUp1 = MockUp(name: "1")
         mockUp1.isMetStub = { _ in true }
@@ -82,7 +87,7 @@ final class SetupLoaderTests: TuistUnitTestCase {
         // given
         let temporaryPath = try self.temporaryPath()
         let projectPath = temporaryPath.appending(component: "Project")
-        try FileHandler.shared.touch(temporaryPath.appending(component: Manifest.setup.fileName(projectPath)))
+        manifestFilesLocator.locateSetupStub = temporaryPath.appending(component: Manifest.setup.fileName(projectPath))
 
         let mockUp1 = MockUp(name: "1")
         mockUp1.isMetStub = { _ in true }
@@ -107,7 +112,7 @@ final class SetupLoaderTests: TuistUnitTestCase {
     func test_meet_when_loadSetup_throws() throws {
         // given
         let projectPath = try temporaryPath()
-        try FileHandler.shared.touch(projectPath.appending(component: Manifest.setup.fileName(projectPath)))
+        manifestFilesLocator.locateSetupStub = projectPath.appending(component: Manifest.setup.fileName(projectPath))
         manifestLoader.loadSetupStub = { _ in throw ManifestLoaderError.manifestNotFound(.setup, projectPath) }
 
         // when / then